move generate base context into context
