+++ butler_offline/viewcore/context/__init__.py
@@ -2,6 +2,8 @@
 '''
 New implementation
 '''
+from butler_offline.viewcore.menu import get_key_for_name, get_name_from_key, get_menu_list
+from butler_offline.viewcore.state import persisted_state
 
 TRANSACTION_ID_KEY = 'ID'
 ERROR_KEY = '%Errortext'
@@ -65,3 +67,15 @@ def is_error(context):
 
 def get_error_message(context):
     return context[ERROR_KEY]
+
+
+def generate_base_context(pagename):
+    return {
+        'active': get_key_for_name(pagename),
+        'active_page_url': '/' + pagename + '/',
+        'active_name': get_name_from_key(pagename),
+        'element_titel': get_name_from_key(pagename),
+        'menu': get_menu_list(),
+        'nutzername': persisted_state.database_instance().name,
+        'extra_scripts': ''
+    }
